+++ clx_task_management/__manifest__.py
@@ -3,7 +3,7 @@
 # See LICENSE file for full copyright & licensing details.
 {
     'name': 'CLX Task Management',
-    'version': '13.3.0.1.9',
+    'version': '13.3.0.1.10',
     'summary': 'CLX Task Management',
     'sequence': 1,
     'description': """ CLX Task Management """,
@@ -40,6 +40,7 @@
         # 'views/product_category_views.xml',
         'views/sale_subscription_line_views.xml',
         'wizard/task_cancel_warning_wizard_views.xml',
+        'wizard/project_excel_report_views.xml',
         'views/assets.xml'
     ],
     'demo': [